requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
brython==3.9.5
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.5
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.1.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
